Delete old OSMD file and add new imports
