+++ client-jb/src/components/ProgressPlayFile.js
@@ -4,7 +4,7 @@
 import React, { useEffect, useState, useRef, useCallback } from 'react';
 import { HotKeys } from 'react-hotkeys';
 import { useParams, useNavigate } from 'react-router-dom';
-import OpenSheetMusicDisplay from './OpenSheetMusicDisplayTemp';
+import OpenSheetMusicDisplay from './OpenSheetMusicDisplay';
 import ControlBar from './ControlBar.js';
 import { makeAudioStreamer } from './audioStreamer.js';
 import CountDownTimer from './CountDownTimer.js';
